# fix headers bench: use `Vec::reserve` for reserve buf capacity