--- a/benches/Cargo.toml
+++ b/benches/Cargo.toml
@@ -20,7 +20,7 @@ byte_reader        = "3.0.0"
 tokio              = { version = "1.37.0", features = ["full"] }
 tracing            = "0.1.4"
 tracing-subscriber = "0.3.18"
-hashbrown          = { version = "0.14.5", features = ["raw"] }
+hashbrown          = { version = "0.14.5", features = ["raw", "inline-more"] }
 
 [dev-dependencies]
 rand = "0.8.5"