--- a/benches/benches/request_headers.rs
+++ b/benches/benches/request_headers.rs
@@ -15,45 +15,45 @@ use ohkami_benches::request_headers::{
 
 
 fn input() -> Vec<u8> {
-    // let input: &[u8; 819] = test::black_box(b"\
-    //     Accept-Language: fr-CH, fr;q=0.9, en;q=0.8, de;q=0.7, *;q=0.5\r\n\
-    //     Authorization: Bearer dummy-authorization-token-sample\r\n\
-    //     Date: Wed, 21 Oct 2015 07:28:00 GMT\r\n\
-    //     Host: localhost:7777\r\n\
-    //     Origin: localhost:3333\r\n\
-    //     Referer: https://developer.mozilla.org/ja/docs/Web/JavaScript\r\n\
-    //     Referrer-Policy: no-referrer\r\n\
-    //     Via: HTTP/1.1 GWA\r\n\
-    //     User-Agent: Mozilla/5.0 (platform; rv:geckoversion) Gecko/geckotrail Firefox/firefoxversion\r\n\
-    //     Transfer-Encoding: identity\r\n\
-    //     Connection: upgrade\r\n\
-    //     Upgrade: a_protocol/1, example ,another_protocol/2.2\r\n\
-    //     Forwarded: for=192.0.2.60; proto=http; by=203.0.113.43\r\n\
-    //     Upgrade-Insecure-Requests: 1\r\n\
-    //     From: webmaster@example.org\r\n\
-    //     X-MyApp-Data: example-custom-header-value\r\n\
-    //     Some-Custom-Header: strawberry\r\n\
-    //     Expect: 100-continue\r\n\
-    //     Cookie: PHPSESSID=298zf09hf012fh2; csrftoken=u32t4o3tb3gg43; _gat=1\r\n\
-    //     Cache-Control: no-cache\r\n\
-    //     \r\n\
-    // ");
-
-    let input: &[u8; 485] = test::black_box(b"\
+    let input: &[u8; 819] = test::black_box(b"\
+        Accept-Language: fr-CH, fr;q=0.9, en;q=0.8, de;q=0.7, *;q=0.5\r\n\
         Authorization: Bearer dummy-authorization-token-sample\r\n\
         Date: Wed, 21 Oct 2015 07:28:00 GMT\r\n\
         Host: localhost:7777\r\n\
         Origin: localhost:3333\r\n\
+        Referer: https://developer.mozilla.org/ja/docs/Web/JavaScript\r\n\
+        Referrer-Policy: no-referrer\r\n\
+        Via: HTTP/1.1 GWA\r\n\
         User-Agent: Mozilla/5.0 (platform; rv:geckoversion) Gecko/geckotrail Firefox/firefoxversion\r\n\
         Transfer-Encoding: identity\r\n\
-        Connection: Keep-Alive\r\n\
+        Connection: upgrade\r\n\
+        Upgrade: a_protocol/1, example ,another_protocol/2.2\r\n\
+        Forwarded: for=192.0.2.60; proto=http; by=203.0.113.43\r\n\
+        Upgrade-Insecure-Requests: 1\r\n\
         From: webmaster@example.org\r\n\
         X-MyApp-Data: example-custom-header-value\r\n\
         Some-Custom-Header: strawberry\r\n\
+        Expect: 100-continue\r\n\
         Cookie: PHPSESSID=298zf09hf012fh2; csrftoken=u32t4o3tb3gg43; _gat=1\r\n\
         Cache-Control: no-cache\r\n\
         \r\n\
     ");
+
+    // let input: &[u8; 485] = test::black_box(b"\
+    //     Authorization: Bearer dummy-authorization-token-sample\r\n\
+    //     Date: Wed, 21 Oct 2015 07:28:00 GMT\r\n\
+    //     Host: localhost:7777\r\n\
+    //     Origin: localhost:3333\r\n\
+    //     User-Agent: Mozilla/5.0 (platform; rv:geckoversion) Gecko/geckotrail Firefox/firefoxversion\r\n\
+    //     Transfer-Encoding: identity\r\n\
+    //     Connection: Keep-Alive\r\n\
+    //     From: webmaster@example.org\r\n\
+    //     X-MyApp-Data: example-custom-header-value\r\n\
+    //     Some-Custom-Header: strawberry\r\n\
+    //     Cookie: PHPSESSID=298zf09hf012fh2; csrftoken=u32t4o3tb3gg43; _gat=1\r\n\
+    //     Cache-Control: no-cache\r\n\
+    //     \r\n\
+    // ");
 
     // let input: &[u8; 301] = test::black_box(b"\
     //     Authorization: Bearer dummy-authorization-token-sample\r\n\