Fix `response::Headers::_write_to`, for-benchmark wrapper of `write_unchecked_to`, to reserve `buf` capacity by `Vec::reserve` instead of `Vec::reserve_exact`. Benchmark examples:

before
```txt
test write_02_http_crate       ... bench:         118.08 ns/iter (+/- 21.05)
test write_03_fxmap            ... bench:          94.93 ns/iter (+/- 5.96)
test write_03_header_hashbrown ... bench:          94.41 ns/iter (+/- 7.58)
test write_03_ohkami           ... bench:         153.62 ns/iter (+/- 12.09)
test write_04_headermap        ... bench:         105.58 ns/iter (+/- 18.24)
```

after
```txt
test write_02_http_crate       ... bench:         116.04 ns/iter (+/- 2.89)
test write_03_fxmap            ... bench:         100.20 ns/iter (+/- 4.12)
test write_03_header_hashbrown ... bench:          99.91 ns/iter (+/- 2.76)
test write_03_ohkami           ... bench:         117.10 ns/iter (+/- 3.13)
test write_04_headermap        ... bench:         124.93 ns/iter (+/- 3.37)
```

---

and

- Activate `"inline-more"` feature of `hashbrown`
- Introduce size tracking in `header_hashbrown`
- Fix input to the same & larger

for more fairness